Refactor: 프로필 이미지 추상화
- 버튼, 버튼이 아닐 경우 isButton props로 판단
