+++ frontend/src/components/ProfileImage/style.ts
